Refactor all js files to use same ES6 syntax   * Rename files based on functionality   * Move Maps, Photos, Camera and Settings from Login page to the Main view and Settings   * Create Logout button that takes you to login page
