+++ client/App/Components/Main.js
@@ -1,256 +1,106 @@
 var React = require('react-native');
-var api = require('../Utils/api');
-var Dashboard = require('./Dashboard');
-var Camera = require('./Camera');
-var Signup = require('./Signup');
+var Icon = require('react-native-vector-icons/Ionicons');
 var Settings = require('./Settings');
 var MapView = require('./MapView');
-var PhotosView = require('./PhotosView');
+var Camera = require('./Camera');
+
 var {
-  View,
-  Text,
   StyleSheet,
-  TextInput,
-  TouchableHighlight,
-  ActivityIndicatorIOS
+  TabBarIOS,
+  Text,
+  View,
 } = React;
 
 var styles = StyleSheet.create({
-  mainContainer: {
+  tabContent: {
     flex: 1,
-    padding: 30,
-    marginTop: 65,
-    flexDirection: 'column',
-    justifyContent: 'center',
-    backgroundColor: '#34495e'
+    alignItems: 'center',
   },
-  title: {
-    marginTop: 10,
-    marginBottom: 25,
-    fontSize: 35,
-    textAlign: 'center',
-    color: '#e74c3c'
+  tabText: {
+    color: 'white',
+    margin: 50,
   },
-  fieldTitle: {
-    marginTop: 10,
-    marginBottom: 15,
-    fontSize: 25,
-    textAlign: 'center',
-    color: '#fff'
-  },
-  searchInput: {
-    height: 50,
-    padding: 4,
-    marginRight: 5,
-    fontSize: 23,
-    borderWidth: 1,
-    borderColor: 'white',
-    borderRadius: 8,
-    color: 'white'
-  },
-  buttonText: {
-    fontStyle: 'italic',
-    fontSize: 20,
-    color: '#000',
-    alignSelf: 'center'
-  },
-  button: {
-    height: 45,
-    flexDirection: 'row',
-    backgroundColor: 'white',
-    borderColor: 'white',
-    borderWidth: 1,
-    borderRadius: 8,
-    marginBottom: 10,
-    marginTop: 30,
-    alignSelf: 'stretch',
-    justifyContent: 'center'
-  },
-  signup: {
-    marginTop: 20,
-    fontSize: 14,
-    textAlign: 'center',
-    textDecorationLine: 'underline'
-  },
-  err: {
-    textAlign: 'center'
-  }
 });
 
 class Main extends React.Component {
   constructor(props) {
     super(props);
-    this.state = {
-      username: '',
-      password: '',
-      isLoading: false,
-      error: false
+    this.displayName = 'Main';
+    this.statics = {
+      title: '<TabBarIOS>',
+      description: 'Tab-based navigation.'
     };
-  }
-
-  handleUsernameChange (event) {
-    this.setState({
-      username: event.nativeEvent.text
-    });
-  }
-
-  handlePasswordChange (event) {
-    this.setState({
-      password: event.nativeEvent.text
-    });
-  }
-
-  handleSubmit(){ 
-    this.setState({
-      isLoading: true
-    });
-
-    api.login(this.state.username, this.state.password)
-      .then((res) => {
-        if(res.status === 500){
-          this.setState({
-             error: 'User not found',
-             isLoading: false
-           });
-        } else {
-          this.props.navigator.push({
-            title: res.name || 'Select an Option',
-            component: Dashboard,
-            passProps: {userInfo: res}
-          });
-          this.setState({
-            isLoading: false,
-            error: false,
-            username: '',
-            password: ''
-          });
-          }
-        }).catch((err) => {
-           this.setState({
-             error: 'User not found' + err,
-             isLoading: false
-           });
-        });
+    this.state = {
+      selectedTab: undefined,
+      notifCount: 0,
+      presses: 0
     }
+  }
 
   gotoSettings() {
+    this.setState({
+      selectedTab: 'profile'
+    });
     this.props.navigator.push({
       component: Settings
-    });
-    this.setState({
-      isLoading: false,
-      error: false,
-      username: ''
-    });
-  }
-
-  handleRedirect() {
-    this.props.navigator.push({
-      component: Signup
-    });
-    this.setState({
-      isLoading: false,
-      error: false,
-      username: ''
     });
   }
 
   openCamera() {
+    this.setState({
+      selectedTab: 'camera'
+    });
     this.props.navigator.push({
       component: Camera
-    });
-    this.setState({
-      isLoading: false,
-      error: false,
-      username: ''
     });
   }
 
   openMaps() {
+    this.setState({
+      selectedTab: 'map',
+    });
     this.props.navigator.push({
       component: MapView
     });
-    this.setState({
-      isLoading: false,
-      error: false,
-      username: ''
-    });
   }
 
-  openPhotos() {
-    this.props.navigator.push({
-      component: PhotosView
-    });
-    this.setState({
-      isLoading: false,
-      error: false,
-      username: ''
-    });
+  _renderContent(color: string, pageText: string) {
+    return (
+      <View style={[styles.tabContent, {backgroundColor: color}]}>
+        <Text style={styles.tabText}>{pageText}</Text>
+      </View>
+    );
   }
 
   render() {
-    var showErr = (
-      this.state.error ? <Text style={styles.err}> {this.state.error} </Text> : <View></View>
-      );
     return (
-      <View style={styles.mainContainer}>
-      <Text style={styles.title}> Profound Mongoose </Text>
-        <Text style={styles.fieldTitle}> Username </Text>
-        <TextInput
-          style={styles.searchInput}
-          value={this.state.username}
-          onChange={this.handleUsernameChange.bind(this)} />
-        <Text style={styles.fieldTitle}> Password </Text>
-        <TextInput
-          style={styles.searchInput}
-          value={this.state.password}
-          onChange={this.handlePasswordChange.bind(this)} />
-        <TouchableHighlight
-          style={styles.button}
-          onPress={this.handleSubmit.bind(this)}
-          underlayColor='white'>
-          <Text style={styles.buttonText}> Sign in </Text>
-        </TouchableHighlight>
-
-        <TouchableHighlight
-          onPress={this.handleRedirect.bind(this)}
-          underlayColor='#34495e'>
-          <Text style={styles.signup}> Dont have an account? Sign Up!  </Text>
-        </TouchableHighlight>
-
-        <TouchableHighlight
-          onPress={this.gotoSettings.bind(this)}
-          underlayColor='#34495e'>
-          <Text style={styles.signup}> gotoSettings  </Text>
-        </TouchableHighlight>
-
-        <TouchableHighlight
-          onPress={this.openCamera.bind(this)}
-          underlayColor='#34495e'>
-          <Text style={styles.signup}> Link to Camera (will implement Auth Later)  </Text>
-        </TouchableHighlight>
-
-        <TouchableHighlight
-          onPress={this.openMaps.bind(this)}
-          underlayColor='#34495e'>
-          <Text style={styles.signup}> Link to MapView </Text>
-        </TouchableHighlight>
-
-        <TouchableHighlight
-          onPress={this.openPhotos.bind(this)}
-          underlayColor='#34495e'>
-          <Text style={styles.signup}> Link to PhotosView </Text>
-        </TouchableHighlight>
-
-        <ActivityIndicatorIOS
-          animating= {this.state.isLoading}
-          color='#111'
-          size='large' />
-        
-        {showErr}
-
-      </View>
-    )
+      <TabBarIOS>
+        <Icon.TabBarItem
+          title="Map"
+          iconName="map"
+          selectedIconName="map"
+          selected={this.state.selectedTab === 'map'}
+          onPress={this.openMaps.bind(this)}>
+          {this._renderContent('#414A8C', 'Map')}
+        </Icon.TabBarItem>
+        <Icon.TabBarItem
+          title="Camera"
+          iconName="ios-camera-outline"
+          selectedIconName="ios-camera"
+          selected={this.state.selectedTab === 'camera'}
+          onPress={this.openCamera.bind(this)}>
+          {this._renderContent('#414A8C', 'Camera')}
+        </Icon.TabBarItem>
+        <Icon.TabBarItem
+          title="Profile"
+          iconName="ios-person-outline"
+          selectedIconName="ios-person"
+          selected={this.state.selectedTab === 'profile'}
+          onPress={this.gotoSettings.bind(this)} >
+          {this._renderContent('#414A8C', 'Profile')}
+        </Icon.TabBarItem>
+      </TabBarIOS>
+    );
   }
 }
