+++ client/App/Components/PhotosView.js
@@ -1,6 +1,6 @@
-'use strict';
-
 var React = require('react-native');
+var _ = require('lodash');
+
 var {
   AppRegistry,
   StyleSheet,
@@ -11,30 +11,29 @@ var {
   ScrollView
 } = React;
 
-var _ = require('lodash');
 var {width, height} = Dimensions.get('window');
 
 var IMAGE_URLS = _.flatten(_.times(10, () => {return ['https://trello-avatars.s3.amazonaws.com/bbbdc13fc56dfee5dcf47bfcd0c943d8/original.png', 'https://trello-avatars.s3.amazonaws.com/1556bc519bd314815e4c4a664e31be22/original.png', 'https://trello-avatars.s3.amazonaws.com/e1740bc7efd3c7117cf047c7be4d9b92/original.png']}));
 var IMAGES_PER_ROW = 3
 
-var ReactNativeLayouts = React.createClass({
-
-  getInitialState() {
-    return {
+class ReactNativeLayouts extends React.Component{
+  constructor(props) {
+    super(props);
+    this.state = {
       currentScreenWidth: width,
       currentScreenHeight: height
-    }
-  },
+    };
+  }
 
   handleRotation(event) {
-    var layout = event.nativeEvent.layout
+    var layout = event.nativeEvent.layout;
     this.setState({currentScreenWidth: layout.width, currentScreenHeight: layout.height })
-  },
+  }
 
   calculatedSize() {
-    var size = this.state.currentScreenWidth / IMAGES_PER_ROW
-    return {width: size, height: size}
-  },
+    var size = this.state.currentScreenWidth / IMAGES_PER_ROW;
+    return {width: size, height: size};
+  }
 
   renderRow(images) {
     return images.map((uri) => {
@@ -42,7 +41,8 @@ var ReactNativeLayouts = React.createClass({
         <Image style={[styles.image, this.calculatedSize()]} source={{uri: uri}} />
       )
     })
-  },
+  }
+
   renderImagesInGroupsOf(count) {
      return _.chunk(IMAGE_URLS, IMAGES_PER_ROW).map((imagesForRow) => {
        return (
@@ -51,16 +51,16 @@ var ReactNativeLayouts = React.createClass({
          </View>
        )
      })
-   },
+   }
 
-   render: function() {
+   render() {
      return (
-       <ScrollView onLayout={this.handleRotation} contentContainerStyle={styles.scrollView}>
+       <ScrollView onLayout={this.handleRotation.bind(this)} contentContainerStyle={styles.scrollView}>
          {this.renderRow(IMAGE_URLS)}
        </ScrollView>
      );
    }
- });
+ };
 
  var styles = StyleSheet.create({
 
@@ -81,5 +81,3 @@ var ReactNativeLayouts = React.createClass({
  });
  
  module.exports = ReactNativeLayouts;
- 
-
